Add the ability to enable / disable individual signal classes.
Currently SignalLoggers can either listen to all events (SignalLogger>>start) or only to a set of specific events (SignalLogger>>startFor:).

This is useful when the user wants to focus in on a specific class of events, but doesn't elegantly handle the case where the problem space is less understood, so many events may be interesting.

Adding an enabled class instance variable to BeaconSignal with the associated set / get methods allows individual signals to be disabled. This will typically be used to gradually filter out irrelevant signals and allow the user to focus on those of interest.

Fixes: https://github.com/feenkcom/gtoolkit/issues/929