Remove the global enabled flag from BeaconSignal and add BeaconSignal class>>excludingNormalOperationSignals.
The global enabled flag is bad because it is global. #startFor: should be used instead.

BeaconSignal class>>excludingNormalOperationSignals provides a simple way to start a logger that doesn't include the noise of signals that are emitted as part of normal Gt operations, e.g.:

CircularMemoryLogger startFor: BeaconSignal excludingNormalOperationSignals


Issue: feenkcom/gtoolkit#2325